debug hash walk 8

+++ gar/tests/conftest.py
@@ -61,8 +61,7 @@ def tempdirwithfiles(tempdir):
 
 
 @pytest.fixture(scope='module')
-def setup(request):
-    print('\nresources_a_setup')
+def setup(create_users, request):
 
     def func1():
         print("func1")
@@ -79,23 +78,27 @@ def setup(request):
         print('teardown the setup')
     request.addfinalizer(teardown)
 
+@pytest.fixture(scope="session")
 def create_users():
     """ Create 3 users and 2 groups """
-    groups = ["group1", "group2"]
-    users = {"user1": ["group1"], "user2": ["group1", "group2"], "user3": ["group2"]}
-    created_groups = []
-    for g in groups:
-        status = os.system(f"sudo addgroup {g}")
-        if status == 0:
-            created_groups.append(g)
-    created_users = {}
-    for u, g in users.items():
-        status = os.system("sudo useradd -r -m -G {} {}".format(" ".join(u),g))
-        if status == 0:
-            created_users.update({u, g})
-    yield (created_groups, created_users)
-    # clean up
-    for g in created_groups:
-        os.system(f"sudo groupdel {g}")
-    for u,g in created_users.items():
-        os.system("sudo userdel -r {u}")
+    if os.environ['CI']:
+        groups = ["group1", "group2"]
+        users = {"user1": ["group1"], "user2": ["group1", "group2"], "user3": ["group2"]}
+        created_groups = []
+        for g in groups:
+            status = os.system(f"sudo addgroup {g}")
+            if status == 0:
+                created_groups.append(g)
+        created_users = {}
+        for u, g in users.items():
+            status = os.system("sudo useradd -r -m -G {} {}".format(" ".join(u),g))
+            if status == 0:
+                created_users.update({u, g})
+        yield (created_groups, created_users)
+        # clean up
+        for g in created_groups:
+            os.system(f"sudo groupdel {g}")
+        for u,g in created_users.items():
+            os.system("sudo userdel -r {u}")
+    else:
+        yield None
